feat: Move vcard code to their own file
